[SMA-320] remove AI score overall

+++ sma-core-service/src/main/java/com/sma/core/dto/request/job/AddJobScoringCriteriaRequest.java
@@ -15,6 +15,6 @@ import lombok.experimental.FieldDefaults;
 @Builder
 public class AddJobScoringCriteriaRequest {
     Integer criteriaId;
-    Double weight;
+    Float weight;
     Boolean enable;
 }
